Split dom5 up into files by purpose.
