+++ .gitignore
@@ -1,6 +1,2 @@
-package-lock.json
-node_modules
-dom5.js
-dom5.d.ts
-test/dom5_test.js
-test/dom5_test.d.ts
+/node_modules
+/lib
